Do not filter automatically
The before_action was removed from the mixin. It must be manually added to controllers.

This makes it a bit less magic, as well as removes the restriction on having an endpoint named "index".

Controllers will now need to add the following line:
before_action :build_filtered_query, only: :index