Version 0.5.0: Retire Query Classes
This is a breaking change: the mixin for controllers was renamed from Filterameter::DeclarativeControllerFilters to Filterameter::DeclarativeFilters. All references must be updated.

Why retire query classes? The goal of Filterameter is to simplify Rails controllers. Extending the functionality to query classes did not add much value and muddied the purpose of the gem.