Support conditional scopes
Use the arity of the scope to determine wether or not it should receive a value. An inline scope without any arguments is treated as a conditional scope: It will not be run if the parameter value is false.

A class method must be used (instead of an inline scope) to define a scope that takes an argument. (That's the only way to be able to tell it has a single parameter).