Add method build_query_from_filters
New method build_query_from_filters provides an easy way to manually build the query. It is an alternative to the callback.

Additionally, this breaks out the logic to dig the filter parameters out of the controller params variable. Method filter_parameters can be overriden to change how the query parameters are selected.