Handle min_only and max_only range options
The range options min_only and max_only were being used by the filter declaration in two ways: to identify what was declared as well as to determine what filter should be used. The result was that the wrong filter was being built (instead of an attribute filter, there were two min or max filters built).

This change splits the question into pairs of predicate methods. The new methods `min_only?` and `max_only` speak to what was declared. The new methods `minimum_range?` and `maximum_range?` speak to what filter should be used. (The original names of `minimum?` and `maximum?` were removed since they are too vague.)