Enable controller tests for declarations (#124)
* New Errors module

Module Errors will support validations against filters and sorts.

* Add validation to AttributeFilter

* Validate scope filters

- ScopeFilter
- ConditionalScopeFilter

* Break out AttributeValidator

The logic from the AttributeFilter can be re-used for the following filters as well:
- ArelFilter
  - MinimumFilter
  - MaximumFilter
- MatchesFilter

* Validate nested filters

* Validate AttributeSort

* Expose parameter names through SubRegistry

This enable both filter and sort names to be accessed through delegations on the registry:
- filter_parameter_names
- sort_parameter_names

* Raise error for multi-argument scope filter

Update filter factory to evaluate scopes based on arity:
- less than one is a conditional scope: this is either an inline scope (which is assumed to be a conditional scope), or a class method with no arguments
- exactly one is a scope
- more than one is an error

If an inline scope has an argument, the issue will be spotted by the validator.

* Add ScopeSort

This sub-classes FilterSort and overrides the validate method to make sure a scope used for sorting takes exactly one argument.

* Controller Validations

This adds method `declarations_validator` to each controller, exposing the class of the same name. The new class provides both `valid?` and `errors` so that each controller can test the declarations.

* Use error classes instead of strings

When validations fail, use error classes to report the failure instead of strings. This provides a catalog of validation errors as well as makes it easy to distinguish declaration validation errors from unexpected errors that may occur.

* Generate validator errors as string