Release 1.0.0: Filterameter!
What if you could increase readability and reduce boilerplate code at the same time as you simplified and sped development of Rails controllers? Sound too good to be true?!? It's not, it's Filterameter.

With the look and feel of a Rails DSL, declare filters and sorts for search pages. The next developer in (hey, maybe that's you!) will thank you. It's obvious what filter parameters are available and easy to add more.